Divide Meet functionality into smaller components
